Move test utils to test folder
