+++ packages/node/src/evm/handlers/initialize-provider.test.ts
@@ -1,4 +1,4 @@
-import { mockEthers } from 'src/test-utils/test-utils';
+import { mockEthers } from 'test/utils/utils';
 const checkAuthorizationStatusesMock = jest.fn();
 const getAirnodeParametersAndBlockNumberMock = jest.fn();
 const getTemplatesMock = jest.fn();
