+++ packages/node/src/evm/handlers/process-transactions.test.ts
@@ -1,4 +1,4 @@
-import { mockEthers } from 'src/test-utils';
+import { mockEthers } from 'test/utils';
 const estimateGasWithdrawalMock = jest.fn();
 const failMock = jest.fn();
 const fulfillMock = jest.fn();
