+++ packages/node/src/handlers/start-coordinator.test.ts
@@ -1,4 +1,4 @@
-import { mockEthers } from 'src/test-utils';
+import { mockEthers } from 'test/utils';
 const checkAuthorizationStatusesMock = jest.fn();
 const getAirnodeParametersAndBlockNumberMock = jest.fn();
 const getTemplatesMock = jest.fn();
